Fixed the crash by implementing the original version of the json library again.

+++ src/json/json.h
@@ -167,8 +167,10 @@ namespace Json {
    typedef __int64 Int64;
    typedef unsigned __int64 UInt64;
 #  else // if defined(_MSC_VER) // Other platforms, use long long
-   typedef long int Int64;
-   typedef unsigned long int UInt64;
+//   typedef long int Int64;
+//   typedef unsigned long int UInt64;
+   typedef long long int Int64;
+   typedef unsigned long long int UInt64;
 #  endif // if defined(_MSC_VER)
    typedef Int64 LargestInt;
    typedef UInt64 LargestUInt;
